Bug Fix / Windows 7, 10 Compatible

+++ Opencv_Template_Initializer/WizardHandler.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml;
-using System.Text.RegularExpressions;
 using System.Windows.Controls;
 
 namespace Opencv_Template_Initializer {
@@ -25,12 +24,13 @@ namespace Opencv_Template_Initializer {
         String WINDOWS_TARGET_PLATFORM;
 
         String CV_PATH;
-        String CV_TOOLS_VER;
         String CV_PATH_LIB_X64 = @"{0}\build\x64\{1}\lib;%(AdditionalLibraryDirectories)";
         String CV_PATH_LIB_X86 = @"{0}\build\x86\{1}\lib;%(AdditionalLibraryDirectories)";
         String CV_PATH_HEADER = @"{0}\build\include;%(AdditionalIncludeDirectories)";
         String CV_LIB_NAME = "opencv_world{0}";
         String CV_LIB_NAME_ENDS_WITH = ".lib;%(AdditionalDependencies)";
+
+        String CV_PATH_VS;
 
         String archi_x64 = "x64";
         String archi_x86 = "Win32";
@@ -44,58 +44,33 @@ namespace Opencv_Template_Initializer {
         String XML_FILE_DOC1 = @"templates\MyTemplate.vstemplate";
         String XML_FILE_DOC2 = @"templates\OpenCV_Template.vcxproj";
 
+        
 
-        public WizardHandler(float osVer, int vsVer, int cvVer, String cv_path) {
+        public WizardHandler(String SDKver, int vsVer, int cvVer, String cv_path) {
 
-            if (osVer <= 6.1f) {
-                WINDOWS_TARGET_PLATFORM = "7";
-            } else if (osVer < 10) {
-                WINDOWS_TARGET_PLATFORM = "8.1";
-
-            } else {
-                WINDOWS_TARGET_PLATFORM = "10.0.10240.0";
-            }
+            WINDOWS_TARGET_PLATFORM = SDKver;
 
             if (vsVer == 141) {
                 PLATFORM_TOOLSET = "v141";
                 TOOLS_VERSION = "15.0";
-                CV_TOOLS_VER = "v15";
+                CV_PATH_VS = "vc15";
             } else {
                 PLATFORM_TOOLSET = "v140";
                 TOOLS_VERSION = "14.0";
-                CV_TOOLS_VER = "v14";
+                CV_PATH_VS = "vc14";
 
             }
 
             CV_PATH = cv_path;
             CV_PATH_LIB_X64 = CV_PATH_LIB_X64.Replace("{0}", cv_path);
+            CV_PATH_LIB_X64 = CV_PATH_LIB_X64.Replace("{1}", CV_PATH_VS);
             CV_PATH_LIB_X86 = CV_PATH_LIB_X86.Replace("{0}", cv_path);
+            CV_PATH_LIB_X86 = CV_PATH_LIB_X86.Replace("{1}", CV_PATH_VS);
+
 
             CV_PATH_HEADER = CV_PATH_HEADER.Replace("{0}", cv_path);
             CV_VER = cvVer.ToString();
             CV_LIB_NAME = CV_LIB_NAME.Replace("{0}", CV_VER);
-
-            //chkOption = _chkOption;
-
-            //string tmp = File.ReadAllText(@"\templates\MyTemplate.vstemplate");
-            //doc1 = new XmlDocument();
-            //doc1.LoadXml("MyTemplate.vstemplate");
-
-            //tmp = File.ReadAllText();
-            try {
-                /*
-                doc1 = new XmlDocument();
-                doc1.Load(@"templates\MyTemplate.vstemplate");
-
-                doc2 = new XmlDocument();
-                doc2.Load(@"templates\OpenCV_Template.vcxproj");
-                */
-            } catch (Exception a) {
-
-            }
-
-            
-
 
         }
 
@@ -105,7 +80,8 @@ namespace Opencv_Template_Initializer {
 
         public void change_windows_path(bool setX64) {
             String win_path = Environment.GetEnvironmentVariable("Path");
-            String new_path = CV_PATH + @"\build\" + (setX64 == true ? @"x64\" : @"x86\") + CV_TOOLS_VER + @"\bin";
+            
+            String new_path = CV_PATH + @"\build\" + (setX64 == true ? @"x64\" : @"x86\") + CV_PATH_VS + @"\bin";
             if (win_path.IndexOf(new_path) >= 0) {
                 // already exist
                 return;
@@ -238,17 +214,40 @@ namespace Opencv_Template_Initializer {
                             elem_ProjectItem.InnerText = key;
                             cursor.AppendChild(elem_ProjectItem);
                         } else {
-                            FileManager.deleteFile(@"templates\" + key);
+                            //FileManager.deleteFile(@"templates\" + key);
                         }
                     }
                 }
 
 
             }
-
-
             doc.Save(XML_FILE_DOC1);
 
+            // --------------------------------------------------------------
+            // --------------------------------------------------------------
+
+            doc.Load(XML_FILE_DOC2);
+            root = doc.DocumentElement;
+
+            foreach (String key in dict.Keys) {
+                if ((bool)dict[key].IsChecked && key!="main.cpp") {
+                    XmlElement elem_ItemGroup = doc.CreateElement("ItemGroup", doc.DocumentElement.NamespaceURI);
+                    XmlElement elem_Media = doc.CreateElement("Media", doc.DocumentElement.NamespaceURI);
+                    elem_Media.SetAttribute("Include", key);
+                    elem_ItemGroup.AppendChild(elem_Media);
+
+                    root.AppendChild(elem_ItemGroup);
+                } else {
+                    if (key == "main.cpp") {
+
+                    } else {
+                        FileManager.deleteFile(@"templates\" + key);
+                    }
+
+                }
+
+            }
+            doc.Save(XML_FILE_DOC2);
         }
 
     }
